separated example table implementation to other file
